+++ tests/__init__.py
@@ -4,5 +4,6 @@
 from .test_base import *
 from .test_containers import *
 from .test_declarations import *
+from .test_fuzzy import *
 from .test_using import *
 from .test_utils import *